pyproject.toml
[tool.poetry]
name = "data-mesh-diag"
version = "0.0.1-alpha"
description = "Data Mesh Diagram Generator"
authors = [
    "Paul Doran <dorzey@gmail.com>"
    ]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/dorzey/data-mesh-diagram-generator.git"

[tool.poetry.scripts]
data_mesh_diag = "data_mesh_diag.main:main"

[tool.poetry.dependencies]
python = "^3.6"
graphviz = "^0.16"
dataclasses = {version = "^0.8", python = "~3.6"}
PyYaml = "^5.4.1"
dataclasses_jsonschema = "^2.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
